wsd: assert on active sessions when unloading
When we have two connections (sessions) and only
one has loaded a view, then it disconnects while
the other session is loading a view, we can end up
unloading with two sessions and two views, which
is a violation of the unloading condition
of having only one session and view.

We now assert on the number of active sessions.

Also, modernize the unit-test.

Change-Id: Ia864e094daf5d794ffdac411d5ea22c98049afa4
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>